refactor InsightsTable name
