+++ frontend/src/scenes/insights/InsightsTable/index.tsx
