--- frontend/src/scenes/insights/TrendLegend/index.ts
@@ -1,1 +1,0 @@
-export * from './InsightsTable'
