+++ frontend/src/scenes/trends/Trends.tsx
@@ -14,7 +14,7 @@ import { ActionsPie, ActionsLineGraph, ActionsBarValueGraph, ActionsTable } from
 import { SaveCohortModal } from './SaveCohortModal'
 import { trendsLogic } from './trendsLogic'
 import { ViewType } from 'scenes/insights/insightLogic'
-import { InsightsTable } from 'scenes/insights/TrendLegend'
+import { InsightsTable } from 'scenes/insights/InsightsTable'
 import { Button } from 'antd'
 
 interface Props {
